+++ src/components/Form/FieldGroup/FieldGroup.test.tsx
@@ -1,7 +1,6 @@
 import { render, screen } from '@testing-library/react'
-import React from 'react'
 
-import FieldGroup from './FieldGroup'
+import { FieldGroup } from './FieldGroup'
 
 describe('FieldGroup tests', () => {
   const props = {
@@ -24,7 +23,7 @@ describe('FieldGroup tests', () => {
       children: 'children',
       errors: {},
       touched: {
-        'field-name': 'error-name'
+        'field-name': true
       }
     }
