--- src/components/Form/FieldGroup/index.js
@@ -1,1 +1,0 @@
-export { default } from './FieldGroup'
